chapter4/infra/requirements.txt
aws-cdk-lib==2.173.0
constructs>=10.0.0,<11.0.0
jinja2
boto3